extracted navbar into components folder
