+++ pages/index.tsx
@@ -7,23 +7,7 @@ import useSound from 'use-sound';
 import { delay } from "framer-motion";
 
 const Rain = dynamic(() => import("../components/Rain"), { ssr: false });
-
-
-function NavBar({ pages }: { pages: string[] }) {
-	return (
-		<nav className="navbar">
-			<ul className="nav-list">
-				{pages.map((page, index) => (
-					<li key={index} className="nav-item">
-						<a href="#" className="nav-link" data-text={page}>
-							{page}
-						</a>
-					</li>
-				))}
-			</ul>
-		</nav>
-	);
-}
+const NavBar = dynamic(() => import("../components/NavBar"), { ssr: true });
 
 export default function Home() {
 	const [bgImageScale, setBgImageScale] = useState(1);
